pyproject.toml
[tool.poetry]
name = "minesweeper"
version = "0.1.0"
description = "Minesweeper game with automatic player"
authors = ["Agustin Iniguez <f.agustin@gmail.com>"]
readme = "README.md"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.2.1"
pillow = "^11.1.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
